requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
unicodecsv==0.14.1
